+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2020-11-19.01
+-------------
+
+* Add error for "you are only allowed one query at a time".
+
 2020-06-17.01
 -------------